+++ docs/CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 3.43.1 (June 1, 2024)
+Version 3.43.1 (June 3, 2024)
 -----------------------------
 
 **User-visible changes:**
@@ -7,6 +7,9 @@ Version 3.43.1 (June 1, 2024)
 
 Renamed `CFAbstractStore#methodValues` to
 `CFAbstractStore#methodCallExpressions`.
+
+Deprecated `isUnassignableByOtherCode()` in favor of `isAssignableByOtherCode()`.
+Deprecated `isUnmodifiableByOtherCode()` in favor of `isModifiableByOtherCode()`.
 
 **Closed issues:**